move Export Modified Video State out of Export Modified video
